bootloader: assorted fixes and cleanup for building with MSC (#5321)
Remove the Visual Studio specific mappings of snprintf() and
vsnprintf() to _snprintf() and _vsnprintf(). The support for actual
snprintf() and vsnprintf() is available in VS2015 and later.

Fixes the failing bootloader tests involving paths with lengths
exceeding PATH_MAX.

The problem is that behavior of _snprintf() is not compatible with
that of snprintf(); on overflow, it does not null-terminate the
string, and it returns -1 instead of the number of characters that
would have been written.

* remove the MSC-specific remapping of POSIX function names

Under Visual Studio, the POSIX function names are deprecated in
favor of MSC-specific ones that start with an underscore, and their
use generates C4996 warning.

Instead of working around this by manually remapping the underscored
names to the POSIX ones via preprocessor macros, simply disable this
specific deprecation warning for the whole project using CFLAGS.

* use CFLAGS to define _CRT_SECURE_NO_WARNINGS

Instead of defining _CRT_SECURE_NO_WARNINGS in each source file,
use CFLAGS to set it for the whole bootloader project. This also
silences the warnings in zlib code.

The TODO message about using safe string functions is likely
irrelevant, as those functions are not portable.

* remove unused variables

Silences the corresponding warnings when building with MSC.

* change pkgstart offset type from int to size_t

Primarily to silence the MSC warnings about conversion from size_t
to int, but size_t also makes more sense as the storage type.

This requires an ugly explict cast to long with fseek() calls to
avoid another warning, but we should actually consider replacing
the fseek() calls with fseeko() (or _fseeki64() under MSC).

* document the VS2015 requirement

+++ bootloader/src/pyi_archive.h
@@ -57,7 +57,7 @@ typedef struct _cookie {
 
 typedef struct _archive_status {
     FILE * fp;
-    int    pkgstart;
+    size_t pkgstart;
     TOC *  tocbuff;
     TOC *  tocend;
     COOKIE cookie;
